2023-02-15 [코드정리] file rename, projector remove
