+++ pretrain_rec.py
